[#154] refactor : 타입 파일명 types.ts로 통일
